Rename the fault-ds to mock-fault-ds.
