--- src/datasources/fault-ds/plugin.json
@@ -1,13 +1,0 @@
-{
-  "name": "OpenNMS FM Datasource",
-  "id": "opennms-helm-fm-ds",
-  "type": "datasource",
-
-  "partials": {
-    "config": "datasources/fault-ds/partials/config.html"
-  },
-
-  "metrics": true,
-  "annotations": false
-
-}
