+++ src/datasources/mock-fault-ds/module.js
@@ -2,10 +2,10 @@ import {OpenNMSFMDatasource} from './datasource';
 import {OpenNMSFMDatasourceQueryCtrl} from './query_ctrl';
 
 class OpenNMSFMDatasourceConfigCtrl {}
-OpenNMSFMDatasourceConfigCtrl.templateUrl = 'datasources/fault-ds/partials/config.html';
+OpenNMSFMDatasourceConfigCtrl.templateUrl = 'datasources/mock-fault-ds/partials/config.html';
 
 class OpenNMSFMDatasourceQueryOptionsCtrl {}
-OpenNMSFMDatasourceQueryOptionsCtrl.templateUrl = 'datasources/fault-ds/partials/query.options.html';
+OpenNMSFMDatasourceQueryOptionsCtrl.templateUrl = 'datasources/mock-fault-ds/partials/query.options.html';
 
 export {
   OpenNMSFMDatasource as Datasource,
