+++ src/datasources/mock-fault-ds/partials/query.editor.html
@@ -1,0 +1,7 @@
+<query-editor-row query-ctrl="ctrl" can-collapse="false">
+  <div class="gf-form-inline">
+    <div class="gf-form">
+      <label class="gf-form-label">OpenNMS Mock FM Datasource</label>
+    </div>
+  </div>
+</query-editor-row>
